fixed data type error and moved merging of east/ west germany to garden step
