+++ dag/health.yml
@@ -641,6 +641,7 @@ steps:
     - snapshot://smoking/2024-05-30/cigarette_sales.zip
   data://garden/smoking/2024-05-30/cigarette_sales:
     - data://meadow/smoking/2024-05-30/cigarette_sales
+    - data://garden/demography/2023-03-31/population
   data://grapher/smoking/2024-05-30/cigarette_sales:
     - data://garden/smoking/2024-05-30/cigarette_sales
-    - data://garden/demography/2023-03-31/population
+
