+++ etl/steps/data/garden/smoking/2024-05-30/cigarette_sales.py
@@ -22,6 +22,11 @@ COLS_WITH_DATA = [
     "all_tobacco_products_grams_per_adult_per_day",
 ]
 
+MAN_KEY = "manufactured_cigarettes_per_adult_per_day"
+HAND_KEY = "handrolled_cigarettes_per_adult_per_day"
+TOTAL_KEY = "total_cigarettes_per_adult_per_day"
+ALL_KEY = "all_tobacco_products_grams_per_adult_per_day"
+
 
 def standardise_years(df):
     new_df = []
@@ -61,16 +66,47 @@ def standardise_years(df):
     return pd.DataFrame(new_df)
 
 
+def include_split_germany(tb, ds_population):
+    """Include data for Germany 1945-1990 in the table by taking weighted average of East and West Germany data"""
+    germany_tb = tb[tb["country"].isin(["West Germany", "East Germany"])]
+    germany_tb = geo.add_population_to_table(germany_tb, ds_population, interpolate_missing_population=True)
+
+    # calculate share of population for each year
+    added_pop = germany_tb[["year", "population"]].groupby("year").sum().reset_index()
+
+    for idx, row in germany_tb.iterrows():
+        germany_tb.loc[idx, "share_of_population"] = (
+            row["population"] / added_pop[added_pop["year"] == row["year"]]["population"].values[0]
+        )
+    # calculate share of cigarettes per adult for weighted average
+    germany_tb[MAN_KEY] = germany_tb[MAN_KEY] * germany_tb["share_of_population"]
+    germany_tb[HAND_KEY] = germany_tb[HAND_KEY] * germany_tb["share_of_population"]
+    germany_tb[TOTAL_KEY] = germany_tb[TOTAL_KEY] * germany_tb["share_of_population"]
+    germany_tb[ALL_KEY] = germany_tb[ALL_KEY] * germany_tb["share_of_population"]
+
+    # sum up values for weighted average
+    germany_tb = germany_tb[COLS_WITH_DATA + ["year"]].groupby("year").sum().reset_index()
+    germany_tb["country"] = "Germany"
+
+    return germany_tb
+
+
+def cast_to_float(df, col):
+    df[col] = df[col].astype("Float64")
+    return df
+
+
 def run(dest_dir: str) -> None:
     # Load inputs.
     #
     # Load meadow dataset.
     ds_meadow = paths.load_dataset("cigarette_sales")
+    # load population data
+    ds_population = paths.load_dataset("population")
 
     # Read table from meadow dataset.
     tb = ds_meadow["cigarette_sales"].reset_index()
 
-    #
     # Process data.
 
     # Fix years column (change dtype to integer and expand timeframes)
@@ -91,12 +127,21 @@ def run(dest_dir: str) -> None:
         ]
     )
 
+    # Calculate weighted average for Germany 1950-1990
+    germany_tb = include_split_germany(tb, ds_population)
+    # include for Germany 1950-1990
+    tb = pr.concat([tb, germany_tb])
+
     # reorder columns
     tb = tb[["year", "country"] + COLS_WITH_DATA]
 
     # set 0 to nan and drop missing values
     tb = tb.replace(0, np.nan)
     tb = tb.dropna(how="all", subset=COLS_WITH_DATA)
+
+    # cast columns to float
+    for col in COLS_WITH_DATA:
+        tb = cast_to_float(tb, col)
 
     # harmonize countries
     tb = geo.harmonize_countries(
@@ -105,7 +150,6 @@ def run(dest_dir: str) -> None:
 
     # remove duplicate data (from hidden rows in excel sheet)
     tb = tb.drop_duplicates(subset=["country", "year"])
-
     tb = tb.format(["country", "year"])
 
     # Save outputs.
